% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bintersect}
\alias{bintersect}
\title{Estimates intersection between buffers}
\usage{
bintersect(b)
}
\arguments{
\item{b}{an SpatialPoligon list with at least two buffers to intersect}
}
\description{
Estimates intersection between buffers
}
