% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoints.R
\name{checkpoints}
\alias{checkpoints}
\title{check sampling site data}
\usage{
checkpoints(p)
}
\arguments{
\item{p}{a set of georreferenced dataPoints}
}
\description{
check sampling site data
}
\concept{Sampling sites can be either lat long coordinates, spatialPoints, spatialPointsDataFrame or simple features objects. This function makes an interpretation of these data and deliver a package-readable input}
\keyword{internal}
