% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cropper}
\alias{cropper}
\title{Crop rasters at maximum buffer size}
\usage{
cropper(r, b)
}
\arguments{
\item{r}{a raster as input}

\item{b}{a Spatial* or Raster* object with extent properties}
}
\description{
Crop rasters at maximum buffer size
}
