% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decouple.R
\name{decouple}
\alias{decouple}
\title{Decouple landscape scales}
\usage{
decouple(r, p, b)
}
\arguments{
\item{r}{a raster object. Raster must be projected to an UTM system, otherwise, buffer sizes will be given in map units. NULL by default}

\item{p}{a set of Spatial* or simple_features points from where buffers will be calculated}

\item{b}{a vector of buffer sizes (radius) to be calculated. If a single value is provided the function returns a simple buffer cut.}
}
\description{
The function decouple different scales from a raster file, origin locations and radii sizes.
}
