% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decouple.specific.R
\name{decouple.specific}
\alias{decouple.specific}
\title{Decouple specific scales}
\usage{
decouple.specific(ls = NULL, specific = NULL)
}
\arguments{
\item{ls}{an output of \code{\link[=decouple]{decouple()}} or \code{\link[=nestedscales]{nestedscales()}} functions}

\item{specific}{which scales will be kept independent of the others. If you have a three layer object from the decouple or nestedscale functions, then specific can be 1, 2, 3 or any combination of it as c(1,2) or c(2,3). Using all scales in the ls object will return the same object.}
}
\description{
This function extract specific scales as  given a multilayer raster object
}
\examples{
\dontrun{
decoupled.ls <- decouple(r,p, c(500,1000,2000,3000))
spec  <- decouple.specific(decoupled.ls, c(1,3))
plot(decoupled.ls$p02)
plot(spec$p02)
}
}
\seealso{
\code{\link[=decouple]{decouple()}}, \code{\link[=nestedscales]{nestedscales()}}
}
