% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupler.R
\name{decoupler}
\alias{decoupler}
\title{function for decoupling outter rings}
\usage{
decoupler(r, p, radius, id)
}
\arguments{
\item{r}{a raster as impute.}

\item{p}{a spatialPoint to use as buffer center}

\item{radius}{a vector of radius to decoupling. If the subjected raster is not in ultrametric system, the radius will be then calculated in map units}

\item{id}{identifiers for each of p objects that will be passed to layers' name}
}
\description{
This function is for internal use only
}
\keyword{internal}
