% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{euglossini}
\alias{euglossini}
\title{Euglossini alpha diversity data}
\format{
A data frame with 15 rows and 7 variables:
\describe{
\item{site}{the id of sampling site, character}
\item{LS}{an aditional name for sampling sites, character}
\item{Richness}{number of species in each site, double}
\item{Abundance}{number of specimens in each site, double}
\item{ComSppAb}{abundance of common species, double}
\item{IntSppAb}{abundance of intermediate frequency species, double}
\item{RareSppAb}{abundance of rare species, double}
}
}
\source{
Carneiro et al. (2021)
}
\usage{
euglossini
}
\description{
Euglossini alpha diversity data
}
\details{
This dataset describes different partition of alpha-diversity in orchid bees (Apidae:Euglossini)
}
\references{
Carneiro LS, Aguiar WM, Priante CF, Ribeiro MC, Frantine-Silva W & Gaglianone MC (2021).	Front. Ecol. Evol., 9, 264. \url{https://doi.org/10.3389/fevo.2021.628319}
}
\author{
Carneiro et al. (2021)
}
\keyword{datasets}
