% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metrics.R
\name{extract_metrics}
\alias{extract_metrics}
\title{Extract metrics}
\usage{
extract_metrics(x = NULL, countNA = NULL)
}
\arguments{
\item{x}{a list from the decouple function}

\item{countNA}{whether to count NA values. If it is "no" (the default) the proportion of each class in the raster layer will be calculated by the sum of valid values. Otherwise, it will be proportional to the whole raster sizes. The allowed values correspond to never count ("no"), only if the count is positive ("ifany") and even for zero counts ("always"). There are some "phatological" cases in which two different kinds of NAs will be treated differently. See the rbase table function for details.}
}
\description{
Extract metrics
}
