% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.scales.names}
\alias{get.scales.names}
\title{Get Scales Names}
\usage{
get.scales.names(x)
}
\arguments{
\item{x}{a data.frame generated by \code{\link[=lsm2multifit]{lsm2multifit()}}}
}
\description{
this function grab the names of the scales to be used as predictos in multifit function or any other.
}
\examples{
\dontrun{
dec        <- decouple(r,p,c(1000,2000,3000))
lsmdata    <- calc_lsm(dec, level = "landscape", metric = "shdi")
data.mfit  <- lsm2multifit(lsmdata, biodata=euglossini,
                           level = "landscape", metrics="shdi")
get.scales.names(data.mfit)
}
}
