% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lbuffers}
\alias{lbuffers}
\title{Creates a list of Buffers from a list of SpatialPoints}
\usage{
lbuffers(p, b)
}
\arguments{
\item{p}{a SpatialPointsDataFrame}

\item{b}{a vector of buffer radii in metters to be calculated}
}
\description{
this function was designed to create a list of SpatialPolygons
}
