% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p}
\alias{p}
\title{Euglossini Sampling Sites SpatialDataFrame}
\format{
SpatialDataFrame with 15 sampling sites coordinates in UTM system
\describe{
  \item{id}{required id for sampling sites, character}
  \item{x}{longitudinal coordinates, double}
  \item{y}{latitudinal coordinates, double}
}
}
\source{
Caneiro et al. (2021)
}
\usage{
p
}
\description{
Euglossini Sampling Sites SpatialDataFrame
}
\details{
Code used to create this data:
points.file <- system.file("extdata/pnts.shp", package="landscapeDecoupler")
p <- read_points(points.file, type = "shp")
}
\references{
Carneiro LS, Aguiar WM, Priante CF, Ribeiro MC, Frantine-Silva W & Gaglianone MC (2021).	Front. Ecol. Evol., 9, 264. \url{https://doi.org/10.3389/fevo.2021.628319}
}
\author{
Carneiro et al. (2021)
}
\keyword{datasets}
