% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_points.R
\name{read_points}
\alias{read_points}
\title{read georeferenced points}
\usage{
read_points(x,type, delimiter,id, lat, long)
}
\arguments{
\item{x}{a valid pathfile to the points}

\item{type}{the specific filetype to be readed. Accepted types:
Shape files: "shp"
txt files: "txt"}

\item{delimiter}{a character to delimite the columns of the text file with the coordinates}

\item{id}{Only aplicable if the type is csv or txt.
The name of the field containing sites id}

\item{lat}{Only aplicable if the type is csv or txt.
The name of the field containing sites latitude coordinates}

\item{long}{Only aplicable if the type is csv or txt.
The name of the field containing sites longitude coordinates}
}
\description{
read georeferenced points
}
\examples{
points.file <- system.file("extdata/pnts.shp", package="landscapeDecoupler")
p <- read_points(points.file,type="shp")
plot(landscapeDecoupler::r)
plot(p, add=T)
text(p, labels=p$id)
}
