% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sbuffers}
\alias{sbuffers}
\title{Creates a list of Buffers from SpatialPoints and a buffer list}
\usage{
sbuffers(p, b)
}
\arguments{
\item{p}{a single SpatialPoint object}

\item{b}{a vector of buffers to be calculated}
}
\description{
this function was designed to be used in parallelal
computation receiving data from lbuffers
}
\seealso{
lbuffers
}
