% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.plot.functions.R
\name{stackplot}
\alias{stackplot}
\title{Stack density plot through scales}
\usage{
stackplot(x=NULL, level="class", metric=NULL)
}
\arguments{
\item{x}{a table output from \link{calc_lsm} or \link{extract_metrics}}

\item{level}{the level from which metrics must be retrived.
For the moment, only "landscape" or "class" are accepted.
Users can filter the dataset or change properties of the ggplot
output object}

\item{metric}{a metric of choice to be ploted}
}
\description{
A simple stack density plot function for mertrics through scales
}
\examples{
\dontrun{
ls.nest <- nestedscales(r,p,b=c(1000,2000,3000))
ls.lsm  <- calc_lsm(ls.nest, metric=c("shdi","sidi","pland"))
stackplot(ls.lsm, level="class", "pland")
stackplot(ls.lsm, level="landscape", "shdi")
}
}
