% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide.lsm.R
\name{wide.lsm}
\alias{wide.lsm}
\title{Landscape Metrics to wide format}
\usage{
wide.lsm(x, level = NULL, metric = NULL, class = NULL)
}
\arguments{
\item{x}{a data frame from \code{\link[=calc_lsm]{calc_lsm()}}}

\item{level}{The level to extract the metric; must be "landscape" or "class".}

\item{metric}{The metric to be extracted. Must be present in the calc_lsm output table}

\item{class}{The class to extract the value if level is "class"}
}
\value{
a data frame in wide format with a single value for each scale
}
\description{
this function is designed to convert \code{\link[=calc_lsm]{calc_lsm()}} output to a wide format, in which each scale (e.g. "X1000", "X1500", ...) will be a column with a single metric value associated with.
}
\examples{
multi.ls <- decouple(r, p, c(1000, 2000, 3000))
lsmdata <- calc_lsm(multi.ls, metric="shdi")
head(lsmdata)
wide.lsm(x = lsmdata, level="landscape", metric="shdi")
}
