% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_scales.R
\name{write_scales}
\alias{write_scales}
\title{write scales}
\usage{
write_scales(
  x = NULL,
  buffer.names = NULL,
  format = "GTiff",
  path = NULL,
  overwrite = T
)
}
\arguments{
\item{x}{an output of decouple function or a list of raster files}

\item{buffer.names}{a vector of names same length as buffer scales or number of raster files on the list to be saved. If null, rasters will be saved as layers names.}

\item{format}{one of the valids output in rgdal package or raster native defaults. For details, see raster::writeRaster help.}

\item{path}{a path to save raster. Working directory is the default}

\item{overwrite}{a boolean whether to overwrite files. TRUE by default}
}
\description{
This function saves each uncoupled scale as file
}
\seealso{
decouple, raster::writeRaster, raster::writeFormats
}
\author{
Wilson Frantine-Silva
}
