% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lsm.R
\name{calc_lsm}
\alias{calc_lsm}
\title{Calculate with LandscapeMetrics}
\usage{
calc_lsm(x, level = NULL, metric = NULL, ...)
}
\arguments{
\item{x}{an output from decouple}

\item{level}{one of the levels considered by landscapemetrics}

\item{metric}{one of the metrics accepted by landscapemetrics}

\item{...}{any o other parameter from calulate_lsm}
}
\description{
this function return a long format data frame by reshaping a calling of calculate_lsm from landscapeMetrics
}
