% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decouple.R
\name{decouple}
\alias{decouple}
\title{Decouple landscape scales}
\usage{
decouple(r = NULL, p = NULL, b = NULL, dec.specific = NULL, progress = TRUE)
}
\arguments{
\item{r}{a raster object. Raster must be projected to an UTM system, otherwise, buffer sizes will be given in map units. NULL by default}

\item{p}{a vector of points from where buffers will be calculated}

\item{b}{a vecotr of buffer sizes (radius) to be calculated. If a single value is provided the function returns a simple buffer cut.}

\item{dec.specific}{this parameter controls whether the function will evaluate specific scales sepparetly. If the parameter is seted as c(1,3) and buffers as c(100, 200, 300, 400), the first and third buffers (100, 300) will be decoupled and the second and fourth will be merged in the final list. By default the parameter is setted as null, so all buffers are analyzed as independent unites.}

\item{progress}{a logical whether to show progress or not.}
}
\description{
The function decouple different scales from a raster file from a set of a buffers calculated from provided points and radius sizes.
}
