% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decouple.specific.R
\name{decouple.specific}
\alias{decouple.specific}
\title{Decouple specific scales}
\usage{
decouple.specific(ls, specific)
}
\arguments{
\item{ls}{an output of decouple or nestedscales functions}

\item{specific}{which scales decouple out. If you have decouple four scales, let's say, three (100, 200, 300) in the decouple or nestedscale functions, then specific can be 1, 2, 3 or any combination of it as c(1,2) or c(2,3). Using all scales in the ls object will return the an equal object.}
}
\description{
Decouple specific scales
}
