% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm2multifit.R
\name{lsm2multifit}
\alias{lsm2multifit}
\title{Convert landscapeMetrics to Multifit format}
\usage{
lsm2multifit(lsm = NULL, biodata = NULL, response = NULL, id.col = "site")
}
\arguments{
\item{biodata}{a dataframe with response variables}

\item{lsmdata}{a dataframe with landscape metrics as \code{\link[=calc_lsm]{calc_lsm()}} output}

\item{level}{a character indicating the metrics level is "landscape" or "class"}

\item{bio.id}{the column of the biodata that has the site id as the \code{\link[=calc_lsm]{calc_lsm()}} output}

\item{which.class}{A character. If level is equal to "class", then the user must indicate which class will be extracted. Otherwise, the function will fail.}
}
\description{
this function was designed to merge outputs from \code{\link[=calc_lsm]{calc_lsm()}} and biological data to be used as reponse variables in \code{\link[=multifit]{multifit()}} function. The function expect a single metric for each site which must have the same
}
\examples{
#For "landscape" level metrics
lsmdata <- calc_lsm(decouple(r,p,c(1000,2000,3000)), level = "landscape", metric = "shdi")
biodata <- euglossini
d2multi <- lsm2multifit(biodata, lsmdata, "landscape", bio.id="site")
head(d2multi)
multifit(mod = "lm", multief = colnames(d2multi)[grepl("X", colnames(d2multi))],
       formula = Abundance ~ multief, data = d2multi,
       criterion = "R2", plot_est = FALSE)

#For "class" level metrics
lsmdata <- calc_lsm(decouple(r,p,c(1000,2000,3000)), level = "class", metric = "pland")
biodata <- euglossini
d2multi <- lsm2multifit(biodata, lsmdata, "class", which.class = 3)
head(d2multi)
multifit(mod = "lm", multief = colnames(d2multi)[grepl("X", colnames(d2multi))],
         formula = Abundance ~ multief, data = d2multi,
         criterion = "R2", plot_est = FALSE)
}
\author{
Wilson Frantine-Silva
}
