% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedscales.R
\name{nestedscales}
\alias{nestedscales}
\title{Design Multiscale landscapes}
\usage{
nestedscales(r, p, b)
}
\arguments{
\item{r}{a raster object. Raster must be projected to an UTM system, otherwise, buffer sizes will be given in map units.}

\item{p}{a set of Spatial* or simple_features points from where buffers will be calculated}

\item{b}{a vector of buffer sizes (radii in meters) to be calculated. If a single value is provided the function returns a simple buffer cut.}
}
\description{
The function crop out buffers from a landscape raster file, origin Spatial points and radii sizes.
}
