% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_decoupled.R
\name{plot_decoupled}
\alias{plot_decoupled}
\title{Plot decoupled landscapes}
\usage{
plot_decoupled(
  x,
  cols = 1,
  rows = 4,
  colors = NULL,
  breaks = NULL,
  pallete = "Terrain",
  reverse.pal = T,
  legend = T,
  ...
)
}
\arguments{
\item{x}{a list of raster objects or a raster object}

\item{cols}{the number of coloumns for a composition. Default = 1}

\item{rows}{the number of rows for a composition. Default = 4}

\item{colors}{a vector with hexadecilma or R-colors to be used for plot. If NULL the function will calculate a pallet based on the number of classes from the raster(s)}

\item{breaks}{the conventional breaks from the plot function. These breaks match the values in the raster object and the vector of colors gived by users or functions}

\item{pallete}{a hcl valid pallete (to see options use hcl.pals()). The defualt is "Terrain"}

\item{reverse.pal}{a logical indicating whether to reverse the pallete}

\item{legend}{a logical whether to the pallete}

\item{...}{other valid parameter for the plot function}
}
\description{
Plot decoupled landscapes
}
\keyword{function}
