% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_decoupled.R
\name{plot_decoupled}
\alias{plot_decoupled}
\title{Plot decoupled landscapes}
\usage{
plot_decoupled(
  x,
  cols = NULL,
  rows = NULL,
  colors = NULL,
  breaks = NULL,
  pallete = "Terrain",
  reverse.pal = T,
  legend = T,
  leg.adj = 1.1,
  ...
)
}
\arguments{
\item{x}{a list of raster objects or a raster object}

\item{cols}{the number of coloumns for a composition. Default = 1}

\item{rows}{the number of rows for a composition. Default = 4}

\item{colors}{a vector with hexadecilma or R-colors to be used for plot. If NULL the function will calculate a pallet based on the number of classes from the raster(s)}

\item{breaks}{the conventional breaks from the plot function. These breaks match the values in the raster object and the vector of colors given by users or functions}

\item{pallete}{a hcl valid pallete (to see options use \link{hcl.pals() }). The defualt is "Terrain"}

\item{reverse.pal}{a logical indicating whether to reverse the pallet}

\item{legend}{a logical whether to the pallet}

\item{leg.adj}{a adjustment multiplyer to legend x position. Usefull when changing cols number in plot area. Default is 1, but values between 0.9 and 1.2 should work fine.}

\item{...}{other valid parameter for the plot function}
}
\description{
Plot decoupled landscapes
}
\details{
this function was designed to facilitate the plotting of objects returned by the functions \link{decouple() }, \link{decouple.specific() }, and \link{nestedscales() }. Users can control specificities about different aspects of plotting categorical raster and creating copositions of such plots.
}
\examples{
r <- landscapeDecoupler::r
p <- landscaleDecoupler::p
b <- c(1000,2000)
ls <- decouple(r, p, b)
plot_decoupled(ls$p02)
plot_decoupled(ls[1:2])
plot_decoupled(ls[1:4], cols = 4, rows = 2)
}
\seealso{
\link[raster:plot() ]{raster::plot() }, \link{hcl.pals() }
}
\keyword{function}
