% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicor.R
\name{predicor}
\alias{predicor}
\title{Predictor correllation throught scales}
\usage{
predicor(cor.data)
}
\arguments{
\item{x}{a dataframe generated with the functions \link{calc_lsm} and tranformed with \code{\link[=lsm2wide]{lsm2wide()}} function}
}
\value{
A data frame containing the correlation values between scales
}
\description{
This function analyze the correlation of a predictor across different scales.
}
\examples{
#Coparing correlation behaviour in nested and decouple multiscales:

b <- c(500, 1000, 1500, 2000, 2500, 3000)

nest.ls <- nestedscales(r, p, b)
dec.ls  <- decouple(r, p, b)

#extracting percentage for each class and each scale
nest.lsm <- calc_lsm(nest.ls, metric = "pland")
dec.lsm  <- calc_lsm(dec.ls,  metric = "pland")

#Comparing correlation for the class = 3
nestcor <- predicor(lsm2wide(nest.lsm, 3))
deccor  <- predicor(lsm2wide(dec.lsm,  3))

corcomp    <- data.frame(
  strategy    = factor(c(rep("nested", nrow(nestcor)), rep("decoupled", nrow(deccor))), levels = c("nested","decoupled")),
  scales      = factor(c(nestcor$scales, deccor$scales), levels=nestcor$scales),
  correlation = c(nestcor$correlation, deccor$correlation)
)
library(ggplot2)
ggplot(corcomp, aes(x=scales, y=correlation, fill=strategy))+
  geom_bar(stat="identity", position="dodge")+
  theme(axis.text = element_text(angle=90))

#Now for a metric at the landscape level
nest.lsm <- calc_lsm(nest.ls, metric = "shdi")
dec.lsm  <- calc_lsm(dec.ls,  metric = "shdi")

nestcor <- predicor(lsm2wide(nest.lsm))
deccor  <- predicor(lsm2wide(dec.lsm ))

corcomp    <- data.frame(
  strategy    = factor(c(rep("nested", nrow(nestcor)), rep("decoupled", nrow(deccor))), levels = c("nested","decoupled")),
  scales      = factor(c(nestcor$scales, deccor$scales), levels=nestcor$scales),
  correlation = c(nestcor$correlation, deccor$correlation)
)
library(ggplot2)
ggplot(corcomp, aes(x=scales, y=correlation, fill=strategy))+
  geom_bar(stat="identity", position="dodge")+
  theme(axis.text = element_text(angle=90))
}
